+++ tests/testoscar/Maths/TestClosedInterval.cpp
@@ -47,6 +47,14 @@ TEST(ClosedInterval, normalized_interpolant_at_returns_1_if_equal_to_upper)
     static_assert(ClosedInterval{-3.0f, 7.0f}.normalized_interpolant_at(7.0f) == 1.0f);
 }
 
+TEST(ClosedInterval, normalized_interpolant_at_returns_0_for_any_finite_input_if_lower_equals_upper)
+{
+    // note: this matches `std::lerp`'s inverse behavior
+    for (float v : {-5.0f, 0.0f, 1.0f, 7.0f}) {
+        ASSERT_EQ(ClosedInterval(1.0f, 1.0f).normalized_interpolant_at(v), 0.0f);
+    }
+}
+
 TEST(ClosedInterval, step_size_returns_expected_answers)
 {
     static_assert(ClosedInterval{0.0f, 0.0f}.step_size(0) == 0.0f);
@@ -72,4 +80,31 @@ TEST(ClosedInterval, contains_also_works_for_ints)
     static_assert(ClosedInterval<int>{-1, 1}.contains(0));
     static_assert(not ClosedInterval<int>{0, 1}.contains(-1));
     static_assert(not ClosedInterval<int>{0, 1}.contains(2));
+}
+
+TEST(ClosedInterval, unit_interval_works_for_floats)
+{
+    static_assert(unit_interval<float>() == ClosedInterval{0.0f, 1.0f});
+}
+
+TEST(ClosedInterval, unit_interval_works_for_doubles)
+{
+    static_assert(unit_interval<double>() == ClosedInterval{0.0, 1.0});
+}
+
+TEST(ClosedInterval, bounding_interval_of_for_single_entry_returns_expected_interval)
+{
+    static_assert(bounding_interval_of(7.0f) == ClosedInterval{7.0f, 7.0f});
+}
+
+TEST(ClosedInterval, bounding_interval_of_for_interval_and_single_entry_returns_expected_results)
+{
+    static_assert(bounding_interval_of(ClosedInterval{0.0f, 0.5f}, 1.0f) == ClosedInterval{0.0f, 1.0f});
+    static_assert(bounding_interval_of(ClosedInterval{0.0f, 0.5f}, -1.0f) == ClosedInterval{-1.0f, 0.5f});
+}
+
+TEST(ClosedInterval, bounding_interval_of_for_optional_interval_and_single_value_returns_expected_results)
+{
+    static_assert(bounding_interval_of<float>(std::nullopt, 1.0f) == ClosedInterval{1.0f, 1.0f});
+    static_assert(bounding_interval_of<float>(std::optional<ClosedInterval<float>>{ClosedInterval<float>{0.0f, 1.0f}}, 1.5f) == ClosedInterval{0.0f, 1.5f});
 }